cathaybkAutoTest/requirements.txt
pytest
PyYAML
selenium
pytest-html
